datacamp/requirements.txt
beautifulsoup4~=4.12.2
psycopg2-binary~=2.9.6
pydantic~=1.10.7
regex-spm~=1.0.0
requests~=2.29.0
tqdm~=4.65.0
validators~=0.20.0
wrapt~=1.15.0
